demote LD preload check to warning
the  preload is useless if the header is wrong just change it

  #!/bin/abc
  #!/bin/env abc

it has dangerous side effects GNU apps can't run in the same shell
anymore
I unset LD preload in favour of running GNU apps I'm the same shell. it
is only nested scripts that could  fail and if
they do just fix them it is that easy instead of ruining the shell

so I don't like this block it's better to warn instead

 # back ground

without  LD preload in bionic nested scripts could fail if they are WRONG they should always use bin/env abc for compatibility . in GNU it is possible to use ld.so.preload instead which is presumably  why this check only applies for bionic

in bionic $PREFIX/etc/ld.so.preload is ignored in favour of these files
that are read only. so the only resort is to fix the header which I have
no idea why it is so difficult to understand

/system/etc/ld.config.arm64.txt
/linkerconfig/ld.config.txt