TERMUX_CONTINUE_BUILD -c flag errors after make
annoying since it requires rebuild from scratch especially on device .
there is is still an error where control files are included at the end.
I just gave up and reset. I still think this should be automatic I will
try to fix my fix for this . it should be easy to keep a marker instead
of asking the builder to remember to switch a flag on and off

these errors remain to be solved these files should not be there in the
final packet

20 09:10 ./
drwxr-xr-x 0/0               0 2025-0
3-20 09:10 ./DEBIAN/
-rw-r--r-- 0/0             199 2025-0
3-20 09:10 ./DEBIAN/control
drwxr-xr-x 0/0               0 20